refactor: Moved the http version of a radiostation to its own file
